fix(renderer): resolve XSS vulnerability and stale closures in copy handlers
